Host compute (#83)
Implementation of the compute interfaces using the host as a device. As
host operations are synchronous, may of them are no-ops or trivial
implementations.

Host device interfaces are implemented in the core of the following
reasons reasons:
- The host device is always available although many operations may not
be implemented.
- There may be multiple plugins implementing the same operation in host
(e.g. FFT may be implemented by either MKL or FFTW). Thus, it is
convenient that all host plugins share the same host device
implementation instead of providing their own.

There are no tests implemented for this PR, as I intend to test
interface implementations though a "black-box" testing suite (WIP).

---------

Co-authored-by: Martin Salinas <ssalinasmartin@gmail.com>