Generate tag & GitHub release (#152)
This workflow will generate a tag & a GitHub release when manually
triggered, starting the release process, as this release generation
should itself trigger after it the wheel generation & distribution
system.

Note: The project's version is retrieved using a python library capable
of reading the `pyproject.toml` file to access the version field.