Only add wheel files (#166)
Having nested folders with non-wheel files was causing issues with pip.
The two possible solutions where:
- Manually add a search param for each nested folder
(`--find-links=dist/linux-`, `--find-links=dist/osx-arm64`)
- Remove the nested folders

Since the nested folders only contained binaries for conda, and this
solution is specifically targeted for a local pip installation, and also
it is a temporary solution, i opted for the second solution, which is
cleaner, and simpler.