Remove old assets from development tag up on new release into it (#178)
If not done (as it is right now), when some binaries are deprecated and
lo longer generated into new releases, only the still supported ones are
generated and replace the ones in the development release, leaving the
deprecated ones in place indefinitely.

This PR adds a step that wipes the release assets before pushing the new
ones, thus, leaving a clean and empty release that will inmediately get
populated with the newly produced binaries.