Implemented layout classes for multidim array (#173)
The layout of the multidimensional array is collapsed into linear
memory. In this PR we define/refactor the classes used to represent
this. I recommend to not bother looking at the GitHub diff. Instead I
woud look directly at the `multidimensional` directory in `include/core`
and `tests/src`. Most of the code has been refactored/changed. The
recommended review order is:

- `strided_axis.*` The atom of the multidimensional array. It describes
an axis by is extent (number of elements) and stride (spacing between
consecutive elements in linear memory)
- `strided_layout.*` Collection of `strided_axes` alongside a base
offset. This fully defines the layout of a muldimensional array
- `layout_reference.*` A reference to another layout (e.g.
`strided_layout`) which is re-used in a Copy on Write basis, allowing to
re-use the same layout for multiple arrays. Shares the same interface as
`strided_layout`
- `index.*` index validation utilities
- `slice.*` An slice is a regular access pattern to a multidimensional
array. It can be used to select ranges of an axis. This file defines the
class and some helper functions.
- subscript_tags.* Defines two special tags new_axis_tag and
ellipsis_tag useful when processing subscripts. A subscript is a
selection operation on the layout. We consider index, slice,
new_axis_tag and ellipsis_tag as subscripts.
- `dynamic_subscript.*` A class that can represent any of the above
mentioned subscripts.
- `broadcast.*` Helper functions to perform broadcast operations in a
generic set of layouts/arrays.
https://numpy.org/devdocs//user/basics.broadcasting.html