Explicitly show license identifier in metadata instead of full file (#248)
Before, when running `pip show xmipp4-core`, it printed the whole
LICENSE file, making it hard to find specific package metadata, which
should be inmediately visible.
Besides that, now [PEP
639](https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#license-and-license-files)
suggests to include the SPDX license identifier (which we are using at
the first line of every source file) instead, with the option of also
including files, but I don't think we need that, the license id should
be enough.

This also requires updating `scikit-build-core` to `0.11.*` to enable
support for use of an id string instead of an object in the license
field.