Removed backend base class (#261)
`XXX_backend` and `XXX_manager` interfaces are too heterogeneous in
order to accept a common base class. Thus, removing the `backend`
interface and `basic_service_manager`. To do so, I have re-implemented
`communication_manager` and `device_manager` using the PIMPL idiom to be
similar to `allocator_manager` and `transfer_manager`