Re-defining synchronous behavior (#285)
Defining that when a `device` returns `nullptr` in queue related calls
(`create_device_queue`, `create_device_event`,
`create_device_to_host_event`), it only supports synchronous behavior.
This better fits the rest of the queue-related calls, as passing
`nullptr` as a queue implies synchronous behavior.

This is the case of the cpu device backend, so updating no-op classes.
Also implemented missing tests for this backend.