Created memory_resource_affinity enum (#311)
This enum replaces the `target_placement` enum previously defined in
`device_context.hpp`. In addition, it takes a central role to query
memory resources to a device. This allows to add new items to the enum
without breaking API compatibility in the future. In addition, some
parts of `memory_context` have been separated to `memory_allocator_pool`