Created operation dispatch mechanism for multidim arrays (#298)
The operation dispatcher is the keystone of the framework. It integrates
both `hardware` related and `multidimensional` related components to
allow delegating `operations` into a particular device.

The dispatcher is provided with the operation to be dispatched, output
operands (which may be uninitialized) and input operands and an
execution context containing all necessary information. With this
information:

1. It sanitizes operands according to the provided information,
potentially performing broadcast to the input and deducing the output.
In addition, if any input invariance is not met, an exception is thrown.
2. It allocates the output if necessary following queue semantics.
3. An appropriate kernel is requested to the `kernel_manager`. The
kernel manager will be responsible of selecting the most adequate
provider according to the execution parameters (device, layout,
numerical types...)
4. The selected kernel is executed following the queue semantics.
5. If a queue was provided, the queue is recorded on all operands to
ensure the data remains valid until execution finishes.